pyproject.toml
[project]
name = "kernel-forge"
version = "0.1.0"
description = "Autonomous GPU kernel optimization system"
requires-python = ">=3.12"
dependencies = [
	"click",
	"rich",
	"asyncssh",
	"aiosqlite",
	"pyyaml",
]

[project.optional-dependencies]
dev = [
	"pytest",
	"pytest-asyncio",
	"ruff",
	"mypy",
]

[project.scripts]
kernel-forge = "kernel_forge.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/kernel_forge"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["W191", "E101"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
